test: split save handlers specs
